% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dlog.pollard}
\alias{dlog.pollard}
\title{Pollard's algorithm for discrete logarithm}
\usage{
dlog.pollard(
  a,
  g,
  m,
  starting_point = 1,
  up_bound = 10000,
  method = "mod",
  disp = TRUE
)
}
\arguments{
\item{a}{power}

\item{g}{base}

\item{m}{modular}

\item{starting_point}{starting point of the algorithm, def. set to 1}

\item{up_bound}{upper bound of steps, def. set to 10000}

\item{method}{choose either "mod" or "con", def. MOD}

\item{disp}{choose whether to display stepwise, def. TRUE}
}
\value{
vector: number and modulo or information of fiasco
}
\description{
Performs Pollard's algorithm for discrete algorithm. Allows user to see stepwise solution.
POllard's algorithm is a probability algorithm, meaning it may not yield result.
In arguments there is set upper bound. There are availble two methods of spliitng the set.
One is with respect to 3, denoted by "mod", the other is "con" staying for consequntly,
meaning that there are split with respect to numeration.
Things such as this method are not explained here. I am sorry for possible misunderstanding
due to different names of algorithm established by authors.
}
