% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{elliptic.dlog.sh}
\alias{elliptic.dlog.sh}
\title{Discrete logarithm on curve}
\usage{
elliptic.dlog.sh(P, Q, a, p, disp = TRUE)
}
\arguments{
\item{P}{starting point, the base}

\item{Q}{finishing point, the power}

\item{a}{C = x^3 + ax}

\item{p}{modulus}

\item{disp}{choose whether to display stepwise solution}
}
\value{
number
}
\description{
Calculates discrete logarithm on the elliptic curve using
Shanks's baby-step giant-step algorithm. Looks for the solution
of equation n*P = Q mod Zp with respect to n
}
