% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eulersPhi}
\alias{eulersPhi}
\title{Euler's Phi}
\usage{
eulersPhi(n)
}
\arguments{
\item{n}{argument}
}
\value{
Value of Euler's totient
}
\description{
Calculates Euler's totient function's value for given entry
}
