% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{factor.po.rho}
\alias{factor.po.rho}
\title{Pollard's rho factorisation}
\usage{
factor.po.rho(n, x0 = 62, C = 1, m = 18)
}
\arguments{
\item{n}{number to be checked}

\item{x0}{it is iterised with starting point at x0, def. 62}

\item{C}{there is used function f of x = x^2 + C, def. C = 1}

\item{m}{number of steps}
}
\value{
number or fiasco
}
\description{
Algorithm for factorisation with given parameters. Freely change parameters,
it is a probability algorithm so may or may not find the result, change of parameters
may allow to get results.
}
