% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{factor.po.rho1}
\alias{factor.po.rho1}
\title{Factorisation rho - 1}
\usage{
factor.po.rho1(n, B, A0)
}
\arguments{
\item{n}{number to be factorised}

\item{B}{parameter. With respect to B there are searched numbers, algorithm searches for lower prime divisors}

\item{A0}{starting point}
}
\value{
number or fiasco
}
\description{
Pollard rho-1 algorithm for factorisation, requires given entries for calculating.
}
