% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find.primeroot}
\alias{find.primeroot}
\title{Find first primeroot}
\usage{
find.primeroot(n)
}
\arguments{
\item{n}{argument}
}
\value{
first prime root
}
\description{
Functions looks for the first primeroot for given n
}
