% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.int}
\alias{is.int}
\title{Check if the element is an integer}
\usage{
is.int(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
boolean value
}
\description{
Built in function is.int checks only format of the number, not if it is actually integer.
This function checks if it is really an integer.
}
