% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.prime}
\alias{is.prime}
\title{Check if number is prime}
\usage{
is.prime(n)
}
\arguments{
\item{n}{number to be checked}
}
\value{
boolean value
}
\description{
Checks if given number is prime number with brute force method
}
