% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.spsp}
\alias{is.spsp}
\title{Strongly pseudo-prime}
\usage{
is.spsp(a, n)
}
\arguments{
\item{a}{number to be checked}

\item{n}{modulus}
}
\value{
boolean value
}
\description{
Check whether the number a is strongly pseudo-prime with respect to n
}
