% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mod.lin.eq}
\alias{mod.lin.eq}
\title{Modular linear equation solver}
\usage{
mod.lin.eq(a, b, m, disp = FALSE)
}
\arguments{
\item{a}{multiplicator of x - the unknown}

\item{b}{to be equivalent of a*x}

\item{m}{modulus}

\item{disp}{choose whether to see the stepwise results}
}
\value{
vector: result,modulus
}
\description{
Solves linear equation of the form ax = b mod m. If it does not, function returns FALSE.
User can see stepwise results.
}
