% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prime.proth}
\alias{prime.proth}
\title{Proth's algorithm for prime}
\usage{
prime.proth(n, disp = TRUE)
}
\arguments{
\item{n}{number}

\item{disp}{choose whether to see solution stepwise, def. TRUE}
}
\value{

}
\description{
Checks whether given number is prime number. Algorithm named after Proth.
}
